+++ src/main/java/com/example/surveyapp/domain/ai/chat/presentation/ChatController.java
@@ -3,9 +3,8 @@ package com.example.surveyapp.domain.ai.chat.presentation;
 import com.example.surveyapp.domain.ai.chat.presentation.dto.ChatRequestDto;
 import com.example.surveyapp.domain.ai.chat.presentation.dto.ChatResponseDto;
 import com.example.surveyapp.domain.ai.chat.presentation.dto.IndexRequestDto;
-import com.example.surveyapp.domain.ai.chat.rag.DocumentIndexer;
+import com.example.surveyapp.domain.ai.chat.application.rag.DocumentIndexer;
 import com.example.surveyapp.domain.ai.chat.application.ChatService;
-import com.example.surveyapp.global.response.ApiResponse;
 import com.example.surveyapp.global.security.jwt.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
