+++ src/test/java/com/example/surveyapp/config/testmockbeans/TestMockBeans.java
@@ -7,8 +7,8 @@ import org.springframework.context.annotation.Bean;
 
 import com.example.surveyapp.domain.admin.application.AdminService;
 import com.example.surveyapp.domain.ai.chat.application.ChatService;
-import com.example.surveyapp.domain.ai.chat.rag.DocumentIndexer;
-import com.example.surveyapp.domain.ai.chat.rag.DocumentSearcher;
+import com.example.surveyapp.domain.ai.chat.application.rag.DocumentIndexer;
+import com.example.surveyapp.domain.ai.chat.application.rag.DocumentSearcher;
 import com.example.surveyapp.domain.order.application.OrderService;
 import com.example.surveyapp.domain.point.application.PointService;
 import com.example.surveyapp.domain.product.application.ProductService;
